Draggable distance is now calculatable
Previously the distance was fixed because the starting point was fixed and i couldn't update its value after dragging it. But now I have moved its update down inside the mounted() block to solve this. Now the correct distance can be calculated to the destination regardless of where you drag the source point to.