Updated collision avoidance function
For each segmented coordinate of intended flight, check all other segmented flights coordinates to see if one is within 120(imported function). If a reserved flight has a coordinate within 120m of a intended flight coordinate, check what time both those coordinates are within that distance at(use same index segmented corrdinates and segmented times to find the corresponding time for that coordinate). Used a 5 minute range (math.Abs) of before and after that time to allow for unplanned delays / UAV not travelling as fast as planned.